Basic Project Creation
I created a GM2 project with a very basic movement and collision system. You can modify the variables in the player create event to modify the movement. This is just a start and has basic platformer code, but it is a good place to start. We can now get working!